Added modified gateway startup solution
A modified version of the simulated_device_cloud_upload sample is now in the repo. It removes the user input test to close the gateway, and therefore allows the solution to be run as a background process. This code has been compiled and provided as the bridge executable in the bin directory.